Ripped out tx style into helper fx, look at demo.go as an example of usage
